Weather loading state. Fixed fetching weather
